+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+- Introduced trailing stop order types
 - Added support for paging of account activity data
 - Added support for specifying reported account activity direction as
   well as `until` and `after` times